# functions/docs/openapi.yaml
openapi: 3.0.0
info:
  title: CollabCanvas API
  version: 1.0.0
  description: |
    REST API for CollabCanvas - A collaborative canvas platform for real-time design and AI integration.
    
    ## Authentication
    All endpoints (except `/api/health`) require authentication using Bearer tokens.
    Include your API token in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```
    
    ## Rate Limits
    - Read operations: 200 requests per minute
    - Write operations: 50 requests per minute
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": {
        "message": "Error description",
        "code": "ERROR_CODE"
      }
    }
    ```
  contact:
    name: CollabCanvas Support
    email: support@collabcanvas.example.com
servers:
  - url: https://your-project.web.app/api
    description: Production server
  - url: http://localhost:5001/your-project/us-central1/api
    description: Local development server
...
paths:
...
  /canvases:
    get:
      summary: List canvases
      description: Get all canvases accessible by the authenticated user
      tags: [Canvases]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of canvases
          content:
            application/json:
              schema:
                type: object
                properties:
                  canvases:
                    type: array
                    items:
                      $ref: '#/components/schemas/Canvas'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create canvas
      description: Create a new canvas
      tags: [Canvases]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: My Design Canvas
                description:
                  type: string
                  example: Canvas for project designs
      responses:
        '201':
          description: Canvas created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canvas'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /canvases/{id}:
    get:
      summary: Get canvas
      description: Get a specific canvas by ID
      tags: [Canvases]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CanvasId'
      responses:
        '200':
          description: Canvas details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canvas'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update canvas
      description: Update canvas metadata (owner only)
      tags: [Canvases]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CanvasId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Canvas updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canvas'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Delete canvas
      description: Delete a canvas and all its objects (owner only)
      tags: [Canvases]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CanvasId'
      responses:
        '204':
          description: Canvas deleted
        '403':
          $ref: '#/components/responses/Forbidden'
  /canvases/{id}/snapshot:
    get:
      summary: Get canvas snapshot
      description: Get complete canvas state with all objects
      tags: [Canvases]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CanvasId'
      responses:
        '200':
          description: Canvas snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  canvas:
                    $ref: '#/components/schemas/Canvas'
                  objects:
                    type: array
                    items:
                      $ref: '#/components/schemas/CanvasObject'
                  objectCount:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: API token for external/AI access
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token for user authentication
  parameters:
    CanvasId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Canvas ID
...
  schemas:
    Canvas:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        ownerId:
          type: string
        collaborators:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CanvasObject:
      type: object
      properties:
        id:
          type: string
        canvasId:
          type: string
        type:
          type: string
          enum: [rectangle, circle, text]
        x:
          type: number
          format: float
        y:
          type: number
          format: float
        width:
          type: number
          format: float
        height:
          type: number
          format: float
        fill:
          type: string
          example: "#808080"
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    example: VALIDATION_ERROR
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    example: UNAUTHORIZED
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    example: FORBIDDEN
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    example: NOT_FOUND